Aggregate raw sensor data to a chosen level

Aggregate raw sensor data to a chosen level

aggregate_sensor_data(sensor_data, config, interval_length)

Arguments

sensor_data
data frame for single sensor returned from pull_sensor()

intverval_length
numeric, the interval length in hours. Default is 1. 0.25 indicates 15 minutes.

Value

a data.table with values for volume, occupancy, and speed

date IDate, the given date

interval_bin numeric, the observations interval bin

{measure}.pct_null numeric, the percentage of observations with null values for the given measure

{measure}.sum numeric, the measure's total over the given interval

{measure}.mean numeric, the measure's mean over the given interval

speed numeric, the mean traffic speed over the given interval

Details

Calculating speed

There are 60 scans per second, which means there are 60*60 = 1,800 scans per
30-second interval. The occupancy value in the 30-second interval data
represents the number of scans that were occupied of the 1,800 scans in that
interval.

With 60 scans per second, 60 seconds per minute there are 3,600 scans per minute.
With 3,600 scans per minute, 60 minutes per hour there are 216,000 scans per hour.
To find the number of scans in 15 minutes, we can multiply 0.25 * 216000 = 54,000 scans.