Pull metro sensor configuration

Read MnDOT JSON feed and wrangle into a tidy dataframe containing 20 variables related to sensor configuration. Useful for mapping (contains lat/lons) and calculating performance measures (contains detector_field).

pull_configuration(return_opt = "in_memory")

Arguments

return_opt
an object of class string which indicates how to return the data. "within_dir" will return the data within the directory as a csv entitled "Configuration of Metro Detectors YYYY-MM-DD". "in-memory" will return the data in R, but requires assignment.

Value

dataframe containing 20 variables, including detector_field and lat/lons, for each sensor in MnDOT's metro district

See also

Examples

if (FALSE) {
config <- pull_configuration("in-memory") # Assign to an object
pull_configuration("within_dir") # No assignment necessary
}